requirements/base.txt
Django==5.1.6
python-environ==0.4.54